Reshape data & add halfway point to spectrogram feature loading

+++ Classification/Convolutional Neural Network/setup_feature_pickles.py
@@ -1,10 +1,12 @@
 import numpy as np
 import os
+import os.path
 import cv2
 import pickle
 import tqdm
 from multiprocessing import Pool
 
+# TODO - find better way than down-scaling the images
 
 # Dimensions of spectrogram
 image_startX = 81
@@ -14,6 +16,9 @@ image_endY = 341
 image_sizeX = image_endX - image_startX
 image_sizeY = image_endY - image_startY
 image_size = image_sizeX * image_sizeY
+image_scaled_sizeX = int(image_sizeX / 3)
+image_scaled_sizeY = int(image_sizeY / 3)
+spectrograms_features_path = "feature_pickles\\spectrogram_features_unsorted.pickle"
 
 
 def load_spectrogram(path):
@@ -25,7 +30,7 @@ def load_spectrogram(path):
         data_point = data_point[image_startY:image_endY, image_startX:image_endX]
 
         # Scale the image down for better memory management
-        data_point = cv2.resize(data_point, (0, 0), fx=0.3, fy=0.3)  # ideally temporary until I figure a better way
+        data_point = cv2.resize(data_point, (image_scaled_sizeX, image_scaled_sizeY))
 
         return [os.path.basename(path).split('.')[0], data_point]
     except Exception as e:
@@ -58,11 +63,16 @@ test_labels = []
 
 def setup_data():
     print('Loading spectrograms...')
-    spectrograms = load_spectrograms()
 
-    pickle_out = open("feature_pickles\\spectrogram_features_unsorted.pickle", "wb")
-    pickle.dump(spectrograms, pickle_out)
-    pickle_out.close()
+    # If this process has been run before, continue from the completion of loading in the spectrograms
+    if os.path.isfile(spectrograms_features_path):
+        pickle_in = open(spectrograms_features_path, "rb")
+        spectrograms = pickle.load(pickle_in)
+    else:
+        spectrograms = load_spectrograms()
+        pickle_out = open(spectrograms_features_path, "wb")
+        pickle.dump(spectrograms, pickle_out)
+        pickle_out.close()
 
     spectrograms_ids.extend([x[0] for x in spectrograms])
     spectrograms_features.extend([x[1] for x in spectrograms])
@@ -101,6 +111,19 @@ if __name__ == "__main__":
     for track_id, labels in tqdm.tqdm(test_labels):
         test_features.append(spectrograms_features[spectrograms_ids.index('{:06d}'.format(int(track_id)))])
 
+    # csv
+    allfeatures = []
+    allfeatures.extend(training_features)
+    allfeatures.extend(validation_features)
+    allfeatures.extend(test_features)
+
+    np.savetxt("feature_pickles\\features.csv", [x[1] if x is not None else None for x in allfeatures], delimiter=",", fmt="%s")
+
+    # Reshape the features ready for use with CNN
+    training_features = np.array(training_features).reshape(-1, image_scaled_sizeX, image_scaled_sizeY, 1)
+    validation_features = np.array(validation_features).reshape(-1, image_scaled_sizeX, image_scaled_sizeY, 1)
+    test_features = np.array(test_features).reshape(-1, image_scaled_sizeX, image_scaled_sizeY, 1)
+
     # Save the data
     print('Saving data to pickles...')
 
@@ -115,17 +138,3 @@ if __name__ == "__main__":
     pickle_out = open("feature_pickles\\test_features.pickle", "wb")
     pickle.dump(test_features, pickle_out)
     pickle_out.close()
-
-    # csv
-    allfeatures = []
-    allfeatures.extend(training_features)
-    allfeatures.extend(validation_features)
-    allfeatures.extend(test_features)
-
-    np.savetxt("features.csv", [x[1] if x is not None else None for x in allfeatures], delimiter=",", fmt="%s")
-
-    # Reshape the features ready for use with CNN
-    # training_features = np.array(training_features).reshape(-1, image_sizeX, image_sizeY, 1)
-    # validation_features = np.array(validation_features).reshape(-1, image_sizeX, image_sizeY, 1)
-    # test_features = np.array(test_features).reshape(-1, image_sizeX, image_sizeY, 1)
-
